refactor : Member Entity, DTO 수정
