+++ src/main/java/com/zerobase/oriticket/domain/members/entity/Member.java
@@ -1,7 +1,7 @@
-package com.zerobase.oriticket.members.entity;
+package com.zerobase.oriticket.domain.members.entity;
 
-import com.zerobase.global.constants.MemberStatus;
-import com.zerobase.global.constants.UserRole;
+import com.zerobase.oriticket.domain.members.constants.MemberStatus;
+import com.zerobase.oriticket.domain.members.constants.UserRole;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -16,6 +16,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.ColumnDefault;
 
 
 @Data
@@ -45,9 +46,9 @@ public class Member {
     @Column(nullable = false, length = 30)
     private String phoneNum;
 
-    @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 10)
-    private UserRole roles;
+    @Enumerated(EnumType.STRING)
+    private UserRole roles = UserRole.MEMBER;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 10)
